Turn pet profile into component
Move pet profile and error into component folders
Move their tests in component folder too
Move their styles
Create a page for pet id page that uses petProfile and error components
Move and write some tests for the pet id page
Move some mock statements after imports for consistency with the majority of files
