+++ src/app/owner/pets/[id]/page.module.scss
@@ -1,15 +1,5 @@
-// Styles partially generated with Copilot autocomplete
-
-@import "../../../globals.scss";
-
 .page {
-    display: grid;
-    grid-template-columns: 1fr;
-    grid-template-rows: auto auto;
-
     .top_bar {
-        grid-row-start: 1;
-
         .back_button {
             padding: 0;
 
@@ -19,125 +9,6 @@
 
             &:hover > * {
                 text-decoration: underline;
-            }
-        }
-    }
-
-    .page_content {
-        display: grid;
-        grid-template-columns: 1fr;
-        padding-left: 8px;
-        padding-right: 8px;
-
-        .header {
-            grid-row-start: 1;
-            display: flex;
-            justify-content: flex-start;
-            font-size: $font-large;
-        }
-
-        .pet_info {
-            display: grid;
-            grid-template-columns: 1fr;
-            grid-template-rows: repeat(2, auto);
-            row-gap: 16px;
-            color: var(--mantine-color-gray-6);
-
-            .pet_image {
-                grid-row: 1 / 2;
-                justify-self: center;
-
-                width: 100%;
-                height: auto;
-            }
-
-            .pet_image_container {
-                position: center;
-                border-radius: 8px;
-                overflow: hidden;
-                background-color: $color-light-blue-white;
-                width: 100%;
-                aspect-ratio: 1 / 1;
-            }
-
-            .pet_info_label {
-                font-weight: bold;
-            }
-
-            .pet_details {
-                display: grid;
-                grid-template-columns: 1fr;
-                column-gap: 20px;
-            }
-
-            .pet_details_column {
-                align-self: center;
-                height: auto;
-                font-size: $font-medium;
-
-                ul {
-                    margin: 0;
-                    padding-left: 0;
-                }
-
-                li {
-                    list-style: none;
-                }
-            }
-        }
-
-        .pet_notes {
-            display: grid;
-            grid-template-columns: 1fr;
-            row-gap: 16px;
-
-            .pet_notes_column {
-                -ms-grid-column-span: 1;
-            }
-
-            .diary_entries_list {
-                display: flex;
-                flex-direction: column;
-                gap: 16px;
-            }
-        }
-    }
-}
-
-@media screen and (min-width: $size-tablet) {
-    .page {
-        .page_content {
-            padding-left: 0;
-            padding-right: 0;
-        }
-    }
-}
-
-@media screen and (min-width: 1040px) {
-    .page {
-        .page_content {
-            .pet_info {
-                grid-template-columns: minmax(200px, 250px) auto;
-                grid-template-rows: 1fr;
-                column-gap: 32px;
-
-                .pet_image {
-                    grid-column: 1 / 2;
-                }
-
-                .pet_details {
-                    grid-template-columns: repeat(2, minmax(250px, 300px));
-                }
-
-                .pet_details_column {
-                    height: 150px;
-                }
-            }
-
-            .pet_notes {
-                grid-template-columns: repeat(2, 1fr);
-                grid-template-rows: 1fr;
-                column-gap: 16px;
             }
         }
     }
